+++ castor/stager/DiskCopyForRecallCInt.cpp
@@ -133,7 +133,7 @@ extern "C" {
   //----------------------------------------------------------------------------
   // Cstager_DiskCopyForRecall_gcWeight
   //----------------------------------------------------------------------------
-  int Cstager_DiskCopyForRecall_gcWeight(castor::stager::DiskCopyForRecall* instance, double* var) {
+  int Cstager_DiskCopyForRecall_gcWeight(castor::stager::DiskCopyForRecall* instance, float* var) {
     *var = instance->gcWeight();
     return 0;
   }
@@ -141,7 +141,7 @@ extern "C" {
   //----------------------------------------------------------------------------
   // Cstager_DiskCopyForRecall_setGcWeight
   //----------------------------------------------------------------------------
-  int Cstager_DiskCopyForRecall_setGcWeight(castor::stager::DiskCopyForRecall* instance, double new_var) {
+  int Cstager_DiskCopyForRecall_setGcWeight(castor::stager::DiskCopyForRecall* instance, float new_var) {
     instance->setGcWeight(new_var);
     return 0;
   }
